requirements.txt
streamlit>=1.42
streamlit-extras>=0.4
streamlit-mermaid>=0.3
langgraph>=0.2
langchain-anthropic>=0.2
langchain-core>=0.3
langsmith>=0.1
pandas>=2.0
scipy>=1.11
python-dotenv>=1.0
requests>=2.31
